split only the subnet-prefix assign part
